Add completion block argument to materialize label show/hide method
This commit makes it so the user can now supply a completion block to
the show/hide methods of the materialize label.

A bug was also fixed where the text was updated on the last iteration of
animating when it wasn't suppose to. This was caused by a missing
`return` line.