Fix bug for flicker of text before animation in materialize label
This commit fixes a bug where the text sometimes flashed before
animating into view for the materialize label. This specifically
happened when you set the text on the label and then told it to animate
in (the text would show then disappear as the text faded in).

This bug was occurring because the time between the first animation
frame and setting the text you wanted was long enough such that the
label got updated with the text it was set with. This was fixed by
making sure we set the text of the label to its correct starting string
before it animates.