Refactor RQMaterializeLabel to create super class
This commit refactors the RQMaterializeLabel to have a super class that
can be configured with a pre animation block and animation block. The
pre animation block will be run before any animations and the animation
block is called during each frame update. The animation block is given a
variety of parameters that the user can use to update the label's
attributed string.

The refactoring out of the new RQAnimationLabel will provide a class
that can easily be built on top of to perform various label animations
without having to worry about actually running the animation and
maintaining the label state.